+++ avs/app/build.gradle.kts
@@ -13,7 +13,8 @@ dependencies {
     implementation(project(":avs:client:site"))
     implementation(project(":avs:provider:accountstore-dynamodb"))
     implementation(project(":common:provider:pendingstore-redis"))
-    implementation(project(":module:crypto-demo"))
+    implementation(project(":avs:provider:certificatesigner-demo"))
+    implementation(project(":avs:provider:userlocalizer-demo"))
     implementation(project(":common:app"))
     implementation(libs.dropwizard.core)
     implementation(libs.jakartaValidation.api)
