+++ avs/provider/certificatesigner-demo/build.gradle.kts
@@ -1,0 +1,15 @@
+plugins {
+    `java-library`
+    id("buildlogic.java-conventions")
+    id("buildlogic.json")
+    id("buildlogic.dagger")
+}
+
+dependencies {
+    api(project(":common:provider:signingkey-demo"))
+
+    implementation(project(":avs:spi"))
+    implementation(project(":common:env"))
+
+    testImplementation(project(":avs:spi-testing"))
+}
