+++ avs/provider/userlocalizer-demo/src/main/java/org/example/age/avs/provider/userlocalizer/demo/DemoAvsUserLocalizerModule.java
@@ -1,0 +1,19 @@
+package org.example.age.avs.provider.userlocalizer.demo;
+
+import dagger.Binds;
+import dagger.Module;
+import org.example.age.avs.spi.AvsVerifiedUserLocalizer;
+
+/**
+ * Dagger module that binds {@link AvsVerifiedUserLocalizer}.
+ * <p>
+ * Depends on an unbound {@link AvsLocalizationKeysConfig}}.
+ */
+@Module
+public abstract class DemoAvsUserLocalizerModule {
+
+    @Binds
+    abstract AvsVerifiedUserLocalizer bindAvsVerifiedUserLocalizer(DemoAvsVerifiedUserLocalizer impl);
+
+    DemoAvsUserLocalizerModule() {}
+}
