--- avs/provider/userlocalizer-demo/src/main/java/org/example/age/provider/userlocalizer/demo/DemoAvsUserLocalizerModule.java
@@ -1,19 +1,0 @@
-package org.example.age.provider.userlocalizer.demo;
-
-import dagger.Binds;
-import dagger.Module;
-import org.example.age.avs.spi.AvsVerifiedUserLocalizer;
-
-/**
- * Dagger module that binds {@link AvsVerifiedUserLocalizer}.
- * <p>
- * Depends on an unbound {@link AvsLocalizationKeysConfig}}.
- */
-@Module
-public abstract class DemoAvsUserLocalizerModule {
-
-    @Binds
-    abstract AvsVerifiedUserLocalizer bindAvsVerifiedUserLocalizer(DemoAvsVerifiedUserLocalizer impl);
-
-    DemoAvsUserLocalizerModule() {}
-}
